Conflicts:
	Documentation/Filters/Generated_FilterDocs.qrc
	Source/DREAM3DLib/Common/DataArray.hpp
	Source/DREAM3DLib/IOFilters/INLWriter.cpp
	Source/DREAM3DLib/SurfaceMeshFilters/SourceList.cmake
	Source/Plugins/SurfaceMeshingPlugin/Code/SurfaceMeshing/SMVtkPolyDataWriter.cpp

+++ Source/Plugins/SurfaceMeshingPlugin/Code/SurfaceMeshing/SMVtkPolyDataWriter.cpp
@@ -572,7 +572,7 @@ void SMVtkPolyDataWriter::execute()
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int SMVtkPolyDataWriter::writeBinaryPointData(const std::string &NodesFile, FILE* vtkFile, size_t nNodes, bool conformalMesh)
+int SMVtkPolyDataWriter::writeBinaryPointData(const std::string &NodesFile, FILE* vtkFile, uint32_t nNodes, bool conformalMesh)
 {
   int err = 0;
   unsigned char nodeData[20];
@@ -609,7 +609,7 @@ int SMVtkPolyDataWriter::writeBinaryPointData(const std::string &NodesFile, FILE
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int SMVtkPolyDataWriter::writeASCIIPointData(const std::string &NodesFile, FILE* vtkFile, size_t nNodes, bool conformalMesh)
+int SMVtkPolyDataWriter::writeASCIIPointData(const std::string &NodesFile, FILE* vtkFile, uint32_t nNodes, bool conformalMesh)
 {
   int err = 0;
   unsigned char nodeData[20];
@@ -642,7 +642,7 @@ int SMVtkPolyDataWriter::writeASCIIPointData(const std::string &NodesFile, FILE*
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int SMVtkPolyDataWriter::writeBinaryCellData(const std::string &TrianglesFile, FILE* vtkFile, size_t nTriangles, bool conformalMesh)
+int SMVtkPolyDataWriter::writeBinaryCellData(const std::string &TrianglesFile, FILE* vtkFile, uint32_t nTriangles, bool conformalMesh)
 {
   static const size_t DATA_COUNT = 9;
   int err = 0;
@@ -651,7 +651,7 @@ int SMVtkPolyDataWriter::writeBinaryCellData(const std::string &TrianglesFile, F
   // Open the triangles file for reading
   FILE* triFile = fopen(TrianglesFile.c_str(), "rb");
 
-  size_t triangleCount = nTriangles;
+  uint32_t triangleCount = nTriangles;
   if(false == conformalMesh)
   {
     triangleCount = nTriangles * 2;
@@ -660,7 +660,7 @@ int SMVtkPolyDataWriter::writeBinaryCellData(const std::string &TrianglesFile, F
   std::vector<int> tri_ids(triangleCount);
   // Write the GrainId Data to the file
   fprintf(vtkFile, "\n");
-  fprintf(vtkFile, "CELL_DATA %ld\n", triangleCount);
+  fprintf(vtkFile, "CELL_DATA %u\n", triangleCount);
   fprintf(vtkFile, "SCALARS GrainID int 1\n");
   fprintf(vtkFile, "LOOKUP_TABLE default\n");
   int tData[DATA_COUNT];
@@ -710,14 +710,14 @@ int SMVtkPolyDataWriter::writeBinaryCellData(const std::string &TrianglesFile, F
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-int SMVtkPolyDataWriter::writeASCIICellData(const std::string &TrianglesFile, FILE* vtkFile, size_t nTriangles, bool conformalMesh)
+int SMVtkPolyDataWriter::writeASCIICellData(const std::string &TrianglesFile, FILE* vtkFile, uint32_t nTriangles, bool conformalMesh)
 {
   static const size_t DATA_COUNT = 9;
   size_t nread = 0;
   // Open the triangles file for reading
   FILE* triFile = fopen(TrianglesFile.c_str(), "rb");
   // Write the GrainId Data to the file
-  int triangleCount = nTriangles;
+  uint32_t triangleCount = nTriangles;
   if(false == conformalMesh)
   {
     triangleCount = nTriangles * 2;
@@ -728,7 +728,7 @@ int SMVtkPolyDataWriter::writeASCIICellData(const std::string &TrianglesFile, FI
   fprintf(vtkFile, "LOOKUP_TABLE default\n");
 
   int tData[DATA_COUNT];
-  for (int i = 0; i < nTriangles; i++)
+  for (size_t i = 0; i < nTriangles; i++)
   {
     nread = fread(tData, sizeof(int), DATA_COUNT, triFile);
     if(nread != DATA_COUNT)
